engineer/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==5.0.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
flower==2.0.1
frozenlist==1.4.1
graphviz==0.20.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
humanize==4.9.0
idna==3.6
ipython==8.21.0
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
multidict==6.0.5
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyTelegramBotAPI==4.15.4
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
systems==0.1.0
telebot==0.0.5
telepot==12.7
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
